Remove the toolchain in riscv-tools
The toolchain pointed to by riscv-tools hasn't been maintained for more
than a year now.  This pointer was necessary when we had out-of-tree
ports of the various toolchain components, but now that everything is
upstream toolchains can be obtained via dozens of standard mechanisms.
Having a broken toolchain so publicly visible causes everyone headaches,
so unless someone wants to step up and maintain this I'm going to remove
it.